# Speed up panel2cs2